backend/requirements.txt
python-dotenv
pymysql
mysqlclient
sqlalchemy
sqlalchemy-filters
flask
flask-cors
marshmallow
Flask-JWT-Extended
werkzeug